requirements.txt
# Streamlit UI dependencies
streamlit
amzqr
Pillow

# API dependencies
fastapi
uvicorn[standard]
python-multipart
